requirements.txt
streamlit
numpy
glob2
pandas
opencv-python-headless
Pillow
matplotlib